Fixed the flash writing bug on RP2040 caused by `flash_cs_force`
In #11, a portion of code was stripped from `flash_rpi_pico.c` from Zephyr's code base
and was inserted here and was revised to be compatible with RP2350. Notably, the type of
the argument was changed from enum into bool, resulting in compatiblity issue with
existing RP2040 code that calls this function with an enum.

Fixes zephyrproject-rtos/zephyr#95729
Fixes zephyrproject-rtos/zephyr#96224